+++ src/main/java/dev/lucasvital/auth/login/RefreshTokenRepository.java
@@ -1,0 +1,5 @@
+package dev.lucasvital.auth.login;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {}